refac: move icon to assets directory
